requirements.txt
asttokens==2.4.1
colorama==0.4.6
executing==2.1.0
icecream==2.1.3
loguru==0.7.2
pillow==11.0.0
Pygments==2.18.0
six==1.16.0
toolz==1.0.0
